add exercise 4.21; add delete auth

+++ part4/controllers/blogs.js
@@ -52,7 +52,23 @@ router.put("/:id", async (request, response) => {
 });
 
 router.delete("/:id", async (request, response) => {
-	await Blog.findByIdAndDelete(request.params.id);
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: "token invalid" });
+	}
+
+	const user = await User.findById(decodedToken.id);
+	const blogToDelete = await Blog.findById(request.params.id);
+	if (
+		blogToDelete &&
+		user &&
+		blogToDelete.user.toString() === user.id.toString()
+	) {
+		await Blog.findByIdAndDelete(request.params.id);
+	} else {
+		return response.status(401).json({ error: "unauthorized user" });
+	}
+
 	response.status(204).end();
 });
